Re-enabled transparency for kitty
I also had to ever so slightly change the background color. This is
because `kitty` would apply transparency to everything that is that
color, not only the terminal background. Since I'm also using
`tokyo-night` as the theme for `nvim` and `btop`, there was no other way
to keep the background of those opaque.